+++ CHANGELOG.md
@@ -57,6 +57,8 @@ The versions follow [semantic versioning](https://semver.org).
 
 - `addheader` now recognises many more extensions. Too many to list here.
 
+- Added BibTex comment style.
+
 ### Changed
 
 - The linter output has been very slightly re-ordered to be more internally